+++ CHANGELOG.md
@@ -8,9 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - `--config-override=<path>` command line argument. `<path>` is absolute file path
 - Support for local GPU metadata files `gpu-data.json` and `os-data.json`. They must be in the working directory, and used in special cases where GitHub is a blocked domain
+- Added Scoop environment variable to list for possible 7-Zip installations, see #97
 
 ### Changed
-- Migrated from discontinued .NET Framework to .NET
+- Migrated from discontinued .NET Framework to .NET 7. TNUC will from this point forward require the .NET 7 x86 Runtime
 - Improved error handling for NVIDIA API and metadata repo
 - Improved libary detection code
 - Remove unwanted image tags and URLs from release description